Move ValidationLossCalculator to loss functions
