+++ ms2deepscore/models/loss_functions.py
@@ -1,6 +1,11 @@
 import numpy as np
 import torch
+from matchms.similarity.vector_similarity_functions import cosine_similarity_matrix
 from torch import nn
+
+from ms2deepscore.benchmarking.calculate_scores_for_validation import get_tanimoto_score_between_spectra
+from ms2deepscore.models import SiameseSpectralModel
+from ms2deepscore.models.SiameseSpectralModel import compute_embedding_array
 
 
 def rmse_loss(outputs, targets):
@@ -109,3 +114,32 @@ def bin_dependent_losses(predictions,
             loss = criterion(true_values[idx], predictions[idx])
             losses[loss_type].append(loss)
     return bin_content, bounds, losses
+
+
+class ValidationLossCalculator:
+    def __init__(self,
+                 val_spectrums,
+                 score_bins):
+        # Check if the spectra only are unique inchikeys
+        inchikeys_list = [s.get("inchikey") for s in val_spectrums]
+        assert len(set(inchikeys_list)) == len(val_spectrums), 'Expected 1 spectrum per inchikey'
+        self.target_scores = get_tanimoto_score_between_spectra(val_spectrums, val_spectrums)
+        self.val_spectrums = val_spectrums
+        self.score_bins = score_bins
+
+    def compute_binned_validation_loss(self,
+                                       model: SiameseSpectralModel,
+                                       loss_types):
+        """Benchmark the model against a validation set.
+        """
+        embeddings = compute_embedding_array(model, self.val_spectrums)
+        ms2ds_scores = cosine_similarity_matrix(embeddings, embeddings)
+        losses = bin_dependent_losses(ms2ds_scores,
+                                      self.target_scores,
+                                      self.score_bins,
+                                      loss_types=loss_types
+                                      )
+        average_losses = {}
+        for loss_type in losses:
+            average_losses[loss_type] = np.mean(losses[loss_type])
+        return average_losses
